% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerClassifRandomForest.R
\docType{data}
\name{LearnerClassifRandomForest}
\alias{LearnerClassifRandomForest}
\alias{mlr_learners_classif.randomForest}
\title{Classification Random Forest Learner}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{LearnerClassif}.}
\usage{
LearnerClassifRandomForest
}
\description{
A \link{LearnerClassif} for a classification random forest implemented in randomForest::randomForest()] in package \CRANpkg{randomForest}.
}
\references{
Breiman, L. (2001).
Random Forests
Machine Learning
\url{https://doi.org/10.1023/A:1010933404324}
}
\keyword{datasets}
