% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learnertemplate-package}
\alias{mlr3learnertemplate}
\alias{mlr3learnertemplate-package}
\title{mlr3learnertemplate: Template for Adding New Learners to 'mlr3'}
\description{
Template for adding new learners to 'mlr3'.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/mlr-org/mlr3learnertemplate/issues}
}

}
\author{
\strong{Maintainer}: Michel Lang \email{michellang@gmail.com} (0000-0001-9754-0393)

Authors:
\itemize{
  \item Quay Au \email{quayau@gmail.com} (0000-0002-5252-8902)
}

}
