% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learnertemplate-package}
\alias{mlr3learnertemplate}
\alias{mlr3learnertemplate-package}
\title{mlr3learnertemplate: Template for Adding New Learners to 'mlr3'}
\description{
Template for adding new learners to 'mlr3'.
}
\author{
\strong{Maintainer}: given.name family.name \email{email@email.com}

}
